refactor: align builtins with signature modules
