--- libs/mql-interpreter/src/libs/common/Alert.ts
@@ -1,6 +1,0 @@
-import type { BuiltinFunction } from "./types";
-
-export const Alert: BuiltinFunction = (...args: any[]) => {
-  console.log(...args);
-  return true;
-};
