--- libs/mql-interpreter/src/libs/common/Print.ts
@@ -1,6 +1,0 @@
-import type { BuiltinFunction } from "./types";
-
-export const Print: BuiltinFunction = (...args: any[]) => {
-  console.log(...args);
-  return 0;
-};
