+++ libs/mql-interpreter/src/libs/common/common.ts
@@ -1,40 +1,35 @@
 // Import paths omit extensions for bundler compatibility.
 import type { BuiltinFunction } from "./types";
 import { formatString } from "./format";
-import type { VirtualTerminal } from "../virtualTerminal";
 import { DateTimeValue } from "../../runtime/datetimeValue";
-
-let terminal: VirtualTerminal | null = null;
-export function setTerminal(t: VirtualTerminal | null): void {
-  terminal = t;
-}
-
-export function getTerminal(): VirtualTerminal | null {
-  return terminal;
-}
+import { getTerminal } from "./terminal";
 
 export const Print: BuiltinFunction = (...args: any[]) => {
   const formatted = args.map((a) => (a instanceof DateTimeValue ? a.toString() : a));
-  if (terminal) return terminal.print(...formatted);
+  const term = getTerminal();
+  if (term) return term.print(...formatted);
   console.log(...formatted);
   return 0;
 };
 
 export const Alert: BuiltinFunction = (...args: any[]) => {
-  if (terminal) return terminal.alert(...args);
+  const term = getTerminal();
+  if (term) return term.alert(...args);
   console.log(...args);
   return true;
 };
 
 export const Comment: BuiltinFunction = (...args: any[]) => {
-  if (terminal) return terminal.comment(...args);
+  const term = getTerminal();
+  if (term) return term.comment(...args);
   console.log(...args);
   return 0;
 };
 
 export const PrintFormat: BuiltinFunction = (fmt: string, ...args: any[]) => {
   const text = formatString(fmt, ...args);
-  if (terminal) return terminal.print(text);
+  const term = getTerminal();
+  if (term) return term.print(text);
   console.log(text);
   return 0;
 };
@@ -49,33 +44,9 @@ export const Sleep: BuiltinFunction = (ms: number) => {
   return 0;
 };
 
-export const EventSetTimer: BuiltinFunction = (seconds: number) => {
-  if (terminal) terminal.setTimer(Number(seconds));
-  return 0;
-};
-
-export const EventSetMillisecondTimer: BuiltinFunction = (ms: number) => {
-  if (terminal) terminal.setTimer(Number(ms) / 1000);
-  return 0;
-};
-
-export const EventKillTimer: BuiltinFunction = () => {
-  if (terminal) terminal.killTimer();
-  return 0;
-};
-
-export const EventChartCustom: BuiltinFunction = (
-  id: number,
-  lparam: number,
-  dparam: number,
-  sparam: string
-) => {
-  if (terminal) terminal.queueChartEvent(id, lparam, dparam, sparam);
-  return true;
-};
-
 export const PlaySound: BuiltinFunction = (file: string) => {
-  return terminal ? terminal.playSound(file) : true;
+  const term = getTerminal();
+  return term ? term.playSound(file) : true;
 };
 export const SendMail: BuiltinFunction = (_to: string, _subj: string, _body: string) => true;
 export const SendNotification: BuiltinFunction = (_msg: string) => true;
@@ -100,80 +71,6 @@ export const WebRequest: BuiltinFunction = (
 ) => {
   if (result) result.value = "";
   return -1;
-};
-
-const globalVars: Record<string, { value: number; time: number }> = {};
-
-export const GlobalVariableSet: BuiltinFunction = (name: string, value: number) => {
-  if (terminal) return terminal.setGlobalVariable(name, value);
-  globalVars[name] = { value, time: Math.floor(Date.now() / 1000) };
-  return value;
-};
-
-export const GlobalVariableGet: BuiltinFunction = (name: string) => {
-  if (terminal) return terminal.getGlobalVariable(name);
-  return globalVars[name]?.value ?? 0;
-};
-
-export const GlobalVariableDel: BuiltinFunction = (name: string) => {
-  if (terminal) return terminal.deleteGlobalVariable(name);
-  const existed = name in globalVars;
-  delete globalVars[name];
-  return existed;
-};
-
-export const GlobalVariableCheck: BuiltinFunction = (name: string) => {
-  if (terminal) return terminal.checkGlobalVariable(name);
-  return name in globalVars;
-};
-
-export const GlobalVariableTime: BuiltinFunction = (name: string) => {
-  if (terminal) return terminal.getGlobalVariableTime(name);
-  return globalVars[name]?.time ?? 0;
-};
-
-export const GlobalVariablesDeleteAll: BuiltinFunction = (prefix = "") => {
-  if (terminal) return terminal.deleteAllGlobalVariables(prefix);
-  let count = 0;
-  for (const k of Object.keys(globalVars)) {
-    if (!prefix || k.startsWith(prefix)) {
-      delete globalVars[k];
-      count++;
-    }
-  }
-  return count;
-};
-
-export const GlobalVariablesTotal: BuiltinFunction = () => {
-  if (terminal) return terminal.globalVariablesTotal();
-  return Object.keys(globalVars).length;
-};
-
-export const GlobalVariableName: BuiltinFunction = (index: number) => {
-  if (terminal) return terminal.getGlobalVariableName(index);
-  const names = Object.keys(globalVars);
-  return names[index] ?? "";
-};
-
-export const GlobalVariableTemp: BuiltinFunction = (name: string, value: number) =>
-  GlobalVariableSet(name, value);
-
-export const GlobalVariableSetOnCondition: BuiltinFunction = (
-  name: string,
-  value: number,
-  check: number
-) => {
-  if (terminal) return terminal.setGlobalVariableOnCondition(name, value, check);
-  if (!globalVars[name] || globalVars[name].value === check) {
-    globalVars[name] = { value, time: Math.floor(Date.now() / 1000) };
-    return true;
-  }
-  return false;
-};
-
-export const GlobalVariablesFlush: BuiltinFunction = () => {
-  if (terminal) return terminal.flushGlobalVariables();
-  return 0;
 };
 
 export const TerminalCompany: BuiltinFunction = () => "MetaQuotes Software Corp.";
