+++ libs/mql-interpreter/src/libs/common/files.ts
@@ -1,5 +1,5 @@
 import type { BuiltinFunction } from "./types";
-import { getTerminal } from "./common";
+import { getTerminal } from "./terminal";
 
 export const FileOpen: BuiltinFunction = (name: string, mode: string) => {
   const term = getTerminal();
