--- libs/mql-interpreter/src/libs/common/iATR.ts
@@ -1,3 +1,0 @@
-import type { BuiltinFunction } from "./types";
-
-export const iATR: BuiltinFunction = (..._args: any[]) => 0;
