--- libs/mql-interpreter/src/libs/common/iMA.ts
@@ -1,3 +1,0 @@
-import type { BuiltinFunction } from "./types";
-
-export const iMA: BuiltinFunction = (..._args: any[]) => 0;
