--- libs/mql-interpreter/src/libs/common/iMACD.ts
@@ -1,3 +1,0 @@
-import type { BuiltinFunction } from "./types";
-
-export const iMACD: BuiltinFunction = (..._args: any[]) => 0;
