--- libs/mql-interpreter/src/libs/common/iRSI.ts
@@ -1,3 +1,0 @@
-import type { BuiltinFunction } from "./types";
-
-export const iRSI: BuiltinFunction = (..._args: any[]) => 0;
