+++ libs/mql-interpreter/src/libs/common/registry.ts
@@ -1,10 +1,5 @@
 import type { BuiltinFunction } from "./types";
 import * as AccountInfo from "./account";
-import { iMA } from "./iMA";
-import { iMACD } from "./iMACD";
-import { iRSI } from "./iRSI";
-import { iATR } from "./iATR";
-import { FileOpen, FileReadString, FileWriteString, FileClose } from "./file";
 import {
   IndicatorBuffers,
   SetIndexBuffer,
@@ -16,7 +11,12 @@ import {
   IndicatorSetInteger,
   IndicatorSetString,
   IndicatorShortName,
-} from "./indicator";
+  iMA,
+  iMACD,
+  iRSI,
+  iATR,
+} from "./indicators";
+import { FileOpen, FileReadString, FileWriteString, FileClose } from "./files";
 import {
   ArrayResize,
   ArrayCopy,
@@ -122,7 +122,7 @@ import {
   TimeMonth,
   TimeSeconds,
   TimeYear,
-} from "./datetime";
+} from "./dateandtime";
 import {
   Bars,
   iBars,
@@ -152,10 +152,6 @@ import {
   GetTickCount64,
   GetMicrosecondCount,
   Sleep,
-  EventSetTimer,
-  EventSetMillisecondTimer,
-  EventKillTimer,
-  EventChartCustom,
   PlaySound,
   SendMail,
   SendNotification,
@@ -165,17 +161,6 @@ import {
   DebugBreak,
   MessageBox,
   WebRequest,
-  GlobalVariableSet,
-  GlobalVariableGet,
-  GlobalVariableDel,
-  GlobalVariableCheck,
-  GlobalVariableTime,
-  GlobalVariablesDeleteAll,
-  GlobalVariablesTotal,
-  GlobalVariableName,
-  GlobalVariableTemp,
-  GlobalVariableSetOnCondition,
-  GlobalVariablesFlush,
   TerminalCompany,
   TerminalName,
   TerminalPath,
@@ -187,8 +172,27 @@ import {
   IsTradeAllowed,
   IsTradeContextBusy,
   UninitializeReason,
-  setTerminal,
 } from "./common";
+import {
+  EventSetTimer,
+  EventSetMillisecondTimer,
+  EventKillTimer,
+  EventChartCustom,
+} from "./eventFunctions";
+import {
+  GlobalVariableSet,
+  GlobalVariableGet,
+  GlobalVariableDel,
+  GlobalVariableCheck,
+  GlobalVariableTime,
+  GlobalVariablesDeleteAll,
+  GlobalVariablesTotal,
+  GlobalVariableName,
+  GlobalVariableTemp,
+  GlobalVariableSetOnCondition,
+  GlobalVariablesFlush,
+} from "./globals";
+import { setTerminal } from "./terminal";
 
 export const coreBuiltins: Record<string, BuiltinFunction> = {
   Print,
@@ -375,175 +379,4 @@ export const envBuiltins: Record<string, BuiltinFunction> = {
   IndicatorShortName,
 };
 
-export {
-  Print,
-  Alert,
-  iMA,
-  iMACD,
-  iRSI,
-  iATR,
-  ArrayResize,
-  ArrayCopy,
-  ArraySetAsSeries,
-  ArrayGetAsSeries,
-  ArrayIsSeries,
-  ArrayIsDynamic,
-  ArraySize,
-  ArrayRange,
-  ArrayDimension,
-  ArrayFree,
-  ArrayInitialize,
-  ArrayFill,
-  ArraySort,
-  ArrayMaximum,
-  ArrayMinimum,
-  ArrayBsearch,
-  ArrayCompare,
-  StringTrimLeft,
-  StringTrimRight,
-  StringLen,
-  StringSubstr,
-  StringAdd,
-  StringBufferLen,
-  StringCompare,
-  StringConcatenate,
-  StringFill,
-  StringFind,
-  StringGetChar,
-  StringSetChar,
-  StringInit,
-  StringReplace,
-  StringSplit,
-  StringToLower,
-  StringToUpper,
-  StringGetCharacter,
-  StringSetCharacter,
-  PrintFormat,
-  Comment,
-  GetTickCount,
-  GetTickCount64,
-  GetMicrosecondCount,
-  Sleep,
-  PlaySound,
-  SendMail,
-  SendNotification,
-  SendFTP,
-  TerminalClose,
-  ExpertRemove,
-  DebugBreak,
-  MessageBox,
-  WebRequest,
-  GlobalVariableSet,
-  GlobalVariableGet,
-  GlobalVariableDel,
-  GlobalVariableCheck,
-  GlobalVariableTime,
-  GlobalVariablesDeleteAll,
-  GlobalVariablesTotal,
-  GlobalVariableName,
-  GlobalVariableTemp,
-  GlobalVariableSetOnCondition,
-  GlobalVariablesFlush,
-  TerminalCompany,
-  TerminalName,
-  TerminalPath,
-  IsTesting,
-  IsOptimization,
-  IsVisualMode,
-  IsDemo,
-  IsConnected,
-  IsTradeAllowed,
-  IsTradeContextBusy,
-  UninitializeReason,
-  CharToString,
-  CharToStr,
-  CharArrayToString,
-  StringToCharArray,
-  StringFormat,
-  StringToTime,
-  TimeToString,
-  TimeToStr,
-  NormalizeDouble,
-  DoubleToString,
-  DoubleToStr,
-  IntegerToString,
-  StringToDouble,
-  StrToDouble,
-  StringToInteger,
-  StrToInteger,
-  MathAbs,
-  MathArccos,
-  MathArcsin,
-  MathArctan,
-  MathCeil,
-  MathCos,
-  MathExp,
-  MathFloor,
-  MathLog,
-  MathLog10,
-  MathMax,
-  MathMin,
-  MathMod,
-  MathPow,
-  MathRand,
-  MathRound,
-  MathSin,
-  MathSqrt,
-  MathSrand,
-  MathTan,
-  MathIsValidNumber,
-  Day,
-  DayOfWeek,
-  DayOfYear,
-  Hour,
-  Minute,
-  Month,
-  Seconds,
-  Year,
-  TimeCurrent,
-  TimeLocal,
-  TimeGMT,
-  TimeDaylightSavings,
-  TimeGMTOffset,
-  TimeToStruct,
-  StructToTime,
-  TimeDay,
-  TimeDayOfWeek,
-  TimeDayOfYear,
-  TimeHour,
-  TimeMinute,
-  TimeMonth,
-  TimeSeconds,
-  TimeYear,
-  Bars,
-  iBars,
-  iBarShift,
-  iOpen,
-  iHigh,
-  iLow,
-  iClose,
-  iTime,
-  iVolume,
-  CopyRates,
-  CopyTime,
-  CopyOpen,
-  CopyHigh,
-  CopyLow,
-  CopyClose,
-  CopyTickVolume,
-  SeriesInfoInteger,
-  RefreshRates,
-  ResetLastError,
-  IndicatorBuffers,
-  SetIndexBuffer,
-  SetIndexLabel,
-  SetIndexShift,
-  IndicatorCounted,
-  IndicatorDigits,
-  IndicatorSetDouble,
-  IndicatorSetInteger,
-  IndicatorSetString,
-  IndicatorShortName,
-  AccountInfo,
-  setTerminal,
-};
+export { setTerminal };
