+++ libs/mql-interpreter/src/libs/common/terminal.ts
@@ -1,0 +1,10 @@
+import type { VirtualTerminal } from "../virtualTerminal";
+
+let terminal: VirtualTerminal | null = null;
+export function setTerminal(t: VirtualTerminal | null): void {
+  terminal = t;
+}
+
+export function getTerminal(): VirtualTerminal | null {
+  return terminal;
+}
